Creates simple summary with backend link
